Extract shared fishing module; consolidate to single fishing page
Move duplicated constants (TARGET_SPECIES, SPECIES_CODE_MAP, SPECIES_COLORS,
_MONTHLY_WATER_TEMP, _MONTHLY_CLOUD_COVER, _TECHNIQUES, _TIME_OF_DAY_ADVICE),
utility functions (_estimate_water_temp, _estimate_cloud_cover, _temp_condition,
_parse_stocking, chart builders) and all cached data loaders into
app/components/fishing_data.py as a single source of truth.

Resolve drifted variants in favour of the fishing.py style: plain hyphens
in depth ranges, no emoji labels, double-dash notation.

app/pages/fishing.py becomes the canonical, feature-complete fishing page and
imports all shared symbols from fishing_data. app/app.py is reduced to a
minimal hub entry point that sets page config and welcomes the user.

Closes APR-168.
